+++ pages/_app.tsx
@@ -1,60 +1,30 @@
 import React from 'react';
 import NProgress from 'nprogress';
-import * as Sentry from '@sentry/node';
 import Router from 'next/router';
 import Head from 'next/head';
-import getConfig from 'next/config';
 import { AppProps } from 'next/app';
-import { RewriteFrames } from '@sentry/integrations';
 import '../styles.css';
 
-if (process.env.NEXT_PUBLIC_SENTRY_DSN) {
-  const config = getConfig();
-  const distDir = `${config.serverRuntimeConfig.rootDir}/.next`;
-  Sentry.init({
-    enabled: process.env.NODE_ENV === 'production',
-    integrations: [
-      new RewriteFrames({
-        iteratee: (frame) => {
-          frame.filename = frame.filename.replace(distDir, 'app:///_next');
-          return frame;
-        },
-      }),
-    ],
-    dsn: process.env.NEXT_PUBLIC_SENTRY_DSN,
-  });
-}
+function MyApp({ Component, pageProps }: AppProps) {
+  React.useEffect(() => {
+    Router.events.on('routeChangeStart', () => {
+      NProgress.start();
+    });
 
-Router.events.on('routeChangeStart', () => {
-  NProgress.start();
-});
+    Router.events.on('routeChangeComplete', () => {
+      NProgress.done();
+    });
 
-Router.events.on('routeChangeComplete', () => {
-  NProgress.done();
-});
-
-Router.events.on('routeChangeError', () => {
-  NProgress.done();
-});
-
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
-function MyApp({ Component, pageProps, err }: AppProps) {
+    Router.events.on('routeChangeError', () => {
+      NProgress.done();
+    });
+  }, []);
   return (
     <>
       <Head>
         <title>Plantilla Next.js | Avila Tek</title>
-        <link
-          href="https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@0,600;1,400&display=swap"
-          rel="preload"
-          as="style"
-        />
-        <link
-          href="https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@0,600;1,400&display=swap"
-          rel="stylesheet"
-        />
       </Head>
-      <Component {...pageProps} err={err} />
+      <Component {...pageProps} />
     </>
   );
 }