+++ pages/index.tsx
@@ -1,4 +1,6 @@
 import React from 'react';
+import Image from 'next/image';
+import logo from '../public/logo.png';
 
 function HomePage() {
   return (
@@ -6,10 +8,10 @@ function HomePage() {
       <section className="max-w-screen-md m-auto h-full">
         <article className="">
           <a href="https://avilatek.dev" target="_blank" rel="noreferrer">
-            <img
+            <Image
               className="w-96 h-auto text-center mx-auto"
               alt="Avila Tek Logo"
-              src="https://avilatek.dev/assets/images/logo_white.png"
+              src={logo}
             />
           </a>
           <h1 className="text-2xl text-center text-gray-700 dark:text-gray-100">